requirements/requirements.txt
scikit-learn
funcy
scikit-multilearn
typer
rich
PyQt5
pyyaml
loguru
pytest
pytest-mock
fiftyone